# pkg/docs/swagger.yaml
basePath: /api/v1
definitions:
  response.ErrorResponse:
    properties:
      error:
        type: string
    type: object
info:
  contact: {}
  description: Kubernetes Pod & Node metrics viewer backend
  title: fl-kube-reader-backend API
  version: "1.0"
paths:
...
  /logs/pod:
    get:
      consumes:
      - application/json
      description: Returns container logs for a pod. Supports tailing N lines and
        fetching previous container logs.
      parameters:
      - description: Pod name
        in: query
        name: pod
        required: true
        type: string
      - description: Container name (optional)
        in: query
        name: container
        type: string
      - description: 'Tail last N lines (default: 0 = all)'
        in: query
        name: tail
        type: integer
      - description: Get logs from the previously terminated container
        in: query
        name: previous
        type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: Log payload
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Server error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      summary: Get pod logs
      tags:
      - logs
...
swagger: "2.0"